+++ index.js
@@ -7,6 +7,7 @@ const { program } = require('commander');
 const checkDependencies = require('./helpers/dependency-check');
 const readConfig = require('./helpers/read-config');
 const processDita = require('./helpers/process-dita');
+const installTempDita = require('./helpers/install-dita-ot');
 
 program
     .name('dita-ot-helper')
@@ -27,6 +28,12 @@ program.option(
     false
 );
 
+program.option(
+    '-i --install',
+    'Download and use a temporary instance of DITA OT (e.g., for CI environments)',
+    undefined
+);
+
 program.parse(process.argv);
 
 processConfigFile(confFilePath);
@@ -35,14 +42,35 @@ processConfigFile(confFilePath);
  * Processes a single config file
  * @param {string} configFilePath The path to the JSON config file
  */
-function processConfigFile(configFilePath) {
+async function processConfigFile(configFilePath) {
     console.info('> Checking dependencies');
-    checkDependencies(['dita']);
+    checkDependencies(program['install'] ? [] : ['dita']);
     console.info('> Dependency check complete');
+
+    let ditaInstallation = {
+        clean: () => {},
+        dita: 'dita',
+    };
+
+    if (program['install'] !== undefined) {
+        ditaInstallation = await installTempDita(
+            typeof program['install'] === 'string'
+                ? program['install']
+                : '3.5.1'
+        );
+    }
+
+    console.log(ditaInstallation);
 
     console.info('> Reading config file');
     const config = readConfig(configFilePath);
     console.info('> Processing config file');
-    processDita(configFilePath, config, !program['verbose']);
+    processDita(
+        ditaInstallation.dita,
+        configFilePath,
+        config,
+        !program['verbose']
+    );
     console.info('> Compilation complete.');
+    console.log(await ditaInstallation.clean());
 }